requirements.txt
numpy
torch>=1.2
torchvision
matplotlib
tensorboard
terminaltables
pillow
tqdm
